Main update: Receptionist can view patient's specific information on the table
Small change: Update layout, style for the patient form; add image
